+++ Runtime/Core/ZoneTypesCollection.cs
@@ -1,10 +1,12 @@
+using SimpleMan.Utilities;
 using System.Collections.Generic;
+using System.Linq;
 using UnityEngine;
 
 
 namespace SimpleMan.Zones.Editor
 {
-    public class ZoneTypesCollection : ScriptableObject
+    public sealed class ZoneTypesCollection : ScriptableObject
 	{
 		//------FIELDS
 		[SerializeField]
@@ -18,6 +20,15 @@ namespace SimpleMan.Zones.Editor
         public IReadOnlyList<GameObject> Prefabs
         {
             get => _prefabs;
+        }
+
+
+
+
+        //------METHODS
+        private void OnValidate()
+        {
+            _prefabs = _prefabs.Validate().ToArray();
         }
     }
 }